Add support for git-series information in VCS branch.
This patch adds support for displaying git-series information as branch
information. Git-series (https://github.com/git-series/git-series) is a
sort of "meta vcs" for git, that uses git. It allows tracking revisions
of branches or series. This makes a series a branch... Sorta.
Internally git-series stores its names as refs, which means running `git
branch` will show them with a 'git-series/' prefix, but .git/HEAD will
not hold a reference to them, instead it's in .git/refs/SHEAD.

This patch adds support for parsing that SHEAD file as a branch name,
but only as a fallback if parsing HEAD returned a SHA value, and will
still return the SHA value if the SHEAD doesn't exist[1] (which the
documentation says will not exist if git-series is in a detached state).

[1] https://github.com/git-series/git-series/blob/master/INTERNALS.md